+++ docs/form-render/demos/dependencies.jsx
@@ -1,0 +1,61 @@
+import React from 'react';
+import { Button, message, Input } from 'antd';
+import FormRender, { useForm } from 'form-render';
+
+const { TextArea } = Input;
+
+const Demo = () => {
+  const form = useForm();
+
+  const schema = {
+    type: 'object',
+    properties: {
+      useSelect: {
+        title: '输入框高度',
+        type: 'number',
+      },
+      select2: {
+        title: '输入框',
+        type: 'string',
+        dependencies: ['useSelect'],
+        widget: 'MyTextEditor',
+      },
+    },
+  };
+
+  const onFinish = (data, errors) => {
+    if (errors.length > 0) {
+      message.error(
+        '校验未通过：' + JSON.stringify(errors.map(item => item.name))
+      );
+    } else {
+      message.success('提交成功：' + JSON.stringify(data));
+    }
+  };
+
+  return (
+    <div>
+      <FormRender
+        form={form}
+        schema={schema}
+        widgets={{ MyTextEditor }}
+        onFinish={onFinish}
+      />
+      <Button type="primary" onClick={form.submit}>
+        提交
+      </Button>
+    </div>
+  );
+};
+
+const MyTextEditor = props => {
+  const { addons } = props;
+  console.log(addons.dependValues);
+  let rows;
+  if (addons && addons.dependValues) {
+    rows = addons.dependValues[0] || 2;
+  }
+  return <TextArea rows={rows} />;
+};
+
+export default Demo;