+++ packages/form-render/src/index.d.ts
@@ -11,9 +11,9 @@ export interface FormParams {
   onValidate?: (valid: any) => void;
   showValidate?: boolean;
   /** 数据分析接口，表单展示完成渲染时触发 */
-  logOnMount?: (any) => void;
+  logOnMount?: (stats: any) => void;
   /** 数据分析接口，表单提交成功时触发，获得本次表单填写的总时长 */
-  logOnSubmit?: (any) => void;
+  logOnSubmit?: (stats: any) => void;
 }
 
 export interface ValidateParams {
@@ -71,6 +71,10 @@ export type WatchProperties = {
 export interface FRProps {
   /** 表单 id */
   id?: string | number;
+  /** 表单顶层的className */
+  className?: string;
+  /** 表单顶层的样式 */
+  style?: any;
   /** 表单 schema */
   schema: any;
   /** form单例 */