+++ packages/form-render/src/utils.js
@@ -373,6 +373,41 @@ export function isExpression(func) {
   return false;
 }
 
+export const parseRootValueInSchema = (schema, rootValue) => {
+  const result = clone(schema);
+  if (isObject(schema)) {
+    Object.keys(schema).forEach(key => {
+      const item = schema[key];
+      if (isObject(item)) {
+        result[key] = parseRootValueInSchema(item, rootValue);
+      } else if (typeof item === 'string') {
+        result[key] = parseSingleRootValue(item, rootValue);
+      }
+    });
+  } else {
+    console.error('schema is not an object:', schema);
+  }
+  return result;
+};
+
+// handle rootValue inside List
+export const parseSingleRootValue = (expression, rootValue = {}) => {
+  if (typeof expression === 'string' && expression.indexOf('rootValue') > 0) {
+    const funcBody = expression.substring(2, expression.length - 2);
+    const str = `
+    return ${funcBody.replace(/rootValue/g, JSON.stringify(rootValue))}`;
+
+    try {
+      return Function(str)();
+    } catch (error) {
+      console.error(error, 'expression:', expression, 'rootValue:', rootValue);
+      return null; // 如果计算有错误，return null 最合适
+    }
+  } else {
+    return expression;
+  }
+};
+
 export function parseSingleExpression(func, formData = {}, dataPath) {
   const parentPath = getParentPath(dataPath);
   const parent = getValueByPath(formData, parentPath) || {};
@@ -626,6 +661,7 @@ export const removeEmptyItemFromList = formData => {
     });
   } else if (Array.isArray(formData)) {
     result = formData.filter(item => {
+      if ([false, 0, ''].indexOf(item) > -1) return true;
       if (item && JSON.stringify(item) !== '{}') {
         return true;
       }
@@ -1058,7 +1094,11 @@ const updateSingleSchema = schema => {
     if (schema.rules && schema.rules.length === 0) {
       delete schema.rules;
     }
-    if (JSON.stringify(schema.props) === '{}') {
+    if (
+      typeof schema.props === 'function' ||
+      (isObject(schema.props) && Object.keys(schema.props).length > 0)
+    ) {
+    } else {
       delete schema.props;
     }
     return schema;